requirements-card-vlm.txt
# pip install -r requirements-card-vlm.txt
# Local BLIP-2 tagging (scripts/tag_card_images_vlm.py)
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
accelerate>=0.27.0
